feat: move session related classes under oneid-ecs-core
